Step progress bar added. Code refactored. Success modal added.
