+++ client/package.json
@@ -15,6 +15,8 @@
     "react-icons": "^4.7.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-step-progress": "^1.0.3",
+    "react-step-progress-bar": "^1.0.3",
     "react-token-auth": "^1.1.8",
     "styled-components": "^5.3.6",
     "web-vitals": "^1.1.2"
