+++ client/src/components/PageThree.js
@@ -1,0 +1,10 @@
+import React, {useEffect, useState, useRef} from 'react';
+
+const PageThree = ({ onButtonClick }) => {
+    return(
+        <div>page three</div>
+    )
+}
+
+export default PageThree
+
